+++ examples/basic-node/jest.config.cjs
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = {
+  testEnvironment: 'node',
+  testMatch: ['<rootDir>/commonjs-jest.test.cjs'],
+  transform: {},
+};